mbserver: prevent hang caused by unsolicited return
The issue here is that a return occurred while a request was pending to
a client. The return fulfilled the pending request but the client was
still attempting to fulfill it. While this was happening another request
comes in that is not fulfilled and it is waiting on the pending client
to return so that it can ask it make a request to it. When the pending
client does return though it can't find the original request (because it
is completed and gone) so it doesn't set the flag to go re-check pending
requests. This causes an issue because the other request is waiting for
it to do so.

The solution is to recheck pending requests when a solicited share
returns even if we didn't find the request that the share was associated
with.

A unit test has also been added to catch this scenario.

ITR#461288 : Machine Hung with "Printing from USB drive. DO NOT TOUCH
DRIVE" displayed.

Signed-off-by: Leah Arrington <sarringt@lexmark.com>