Define HAVE_SYSTEMD if we have systemd
configure.ac created a conditional if systemd was being used, but
Makefile.am did not use it. Actually define HAVE_SYSTEMD if we have it.

This fixes a problem where the first few send() calls to
membroker.socket would not be passed on to membroker.service.

Also fix the resulting compile error by using 1 rather than true which
is not defined.