get_kernel_mem_total: include swap pages
In all-mem-except mode, we want to base the calculation on the system's
overall ability to handle dirty pages.  This means we should count swap
pages in addition to normal RAM.  There may be a performance difference
between these pools, but if we want to tune for them differently that
should be handled at a different level.