mbserver: add Request::requested_pages for more accurate logging
It should be as simple as "needed_pages + acquired_pages = how much they
requested."

However, if we get unsolicited pages, we'll knock down the value of
needed_pages in the Request.

This can cause weird things like "got 1024 of 0 reserved pages," when
unsolicited pages fulfill our request.

To fix that, add a requested_pages field, which holds the original
and unmodified request value, so we can log sanely.